core: move svc_cache.h from local directory to generic includes
